refactor: Introduce `FeatureRule` and move `TransformERC20` calldata generation logic into a rule [LIT-760] (#1195)
* Move `getMaxQuoteSlippageRate` to util

* Move `TransformERC20` calldata generation logic to `TransformERC20Rule`
