+++ src/asset-swapper/quote_consumers/feature_rules/types.ts
@@ -1,0 +1,10 @@
+import { CalldataInfo, ExchangeProxyContractOpts, SwapQuote } from '../../types';
+
+export interface FeatureRule {
+    isCompatible(quote: SwapQuote, opts: ExchangeProxyContractOpts): boolean;
+    createCalldata(quote: SwapQuote, opts: ExchangeProxyContractOpts): CalldataInfo;
+}
+
+export interface FeatureRuleRegistry {
+    getTransformErc20Rule(): FeatureRule;
+}
